% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somar-dois-numeros.R
\name{somar_dois_numeros}
\alias{somar_dois_numeros}
\title{soma dois numeros}
\usage{
somar_dois_numeros(x, y)
}
\arguments{
\item{x}{um numero}

\item{y}{outro numero}
}
\value{
um numero
}
\description{
soma dois numeros
}
\examples{
somar_dois_numeros(1, 2)
}
